Fix the Typing problem.
It seems that when you define a specific type in tsconfig.json, it means
you have to explicitly list all of the types your project depends on.
Since this isn't something we are interested in doing, I'm removing that
line again.  I'm opening an issue in the vue-typescript-import-dts repo to
get this published to the DefinitelyTyped repo, but for now if you
manually copy the index.d.ts file to a "vue-typescript-import-dts"
directory under "@Types" then it will work.