First pass at style updates
This is only the first pass, and is not according to redlines yet.  One of
the more significant changes here is to switch over to use the Office
UI Fabric styles.  This is something we're going to need to pull into the
component itself (vs. require it to be included outside).